% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipunen-api.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(resource, limit = NULL)
}
\arguments{
\item{resource}{character name of the resource. Name provided must be a valid
resource name.}

\item{limit}{numeric amount of rows to fetch}
}
\value{
tibble of query parameters.
}
\description{
Low level function used for getting the data for
a given resource endpoint.
}
