% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipunen-api.R
\name{get_parameters}
\alias{get_parameters}
\title{get_parameters}
\usage{
get_parameters(resource)
}
\arguments{
\item{resource}{character name of the resource. Name provided must be a valid
resource name.}
}
\value{
tibble of query parameters.
}
\description{
Low level function used for getting the valid API query parameters for
a given resource endpoint.
}
\examples{
params <- get_parameters("julkaisut")
}
