% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipunen-api.R
\name{get_resource_names}
\alias{get_resource_names}
\title{get_resource_names}
\usage{
get_resource_names()
}
\value{
vector of valid resource names
}
\description{
Get the resource names available through the API.
}
\examples{
get_resource_names()
}
