% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_resource}
\alias{valid_resource}
\title{valid_resource}
\usage{
valid_resource(x)
}
\arguments{
\item{x}{character name of the resource}
}
\value{
logical TRUE/FALSE
}
\description{
Test if a provided argument is a valid resource in Vipunen API. Valid
resources are fetched from the API.
}
\examples{
# TRUE
valid_resource("julkaisut")
# FALSE
valid_resource("foobar")
}
