% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipunen-api.R
\name{vipunen_api}
\alias{vipunen_api}
\title{vipunen_api}
\usage{
vipunen_api(path)
}
\arguments{
\item{path}{character url to be appended to the host.}
}
\value{
resonse object
}
\description{
Return a full path to vipunen API with \code{path} url being appended to the
host path \code{http://api.vipunen.fi}.
}
