% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipunen-cite.R
\name{vipunen_cite}
\alias{vipunen_cite}
\title{Create a Data Bibliography}
\usage{
vipunen_cite(resource, format = "Biblatex")
}
\arguments{
\item{resource}{resource name from Vipunen API}

\item{format}{Default is "Biblatex", alternatives are "bibentry" or "Bibtex".}
}
\value{
a Biblatex, bibentry or Bibtex object.
}
\description{
Crates a bibliography from selected resource.
}
\examples{
\dontrun{
vipunen_cite("koulutusluokitus")
}
}
\references{
See citation("sotkanet")
}
\seealso{
\code{\link[utils:bibentry]{utils::bibentry()}} \code{\link[RefManageR:toBiblatex]{RefManageR::toBiblatex()}}
}
