% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipunen-interactive.R
\name{vipunen_interactive}
\alias{vipunen_interactive}
\title{vipunen_interactive}
\usage{
vipunen_interactive()
}
\value{
fetched data and prompt
}
\description{
Interactive function that helps to fetch data from Vipunen API
}
