% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipunen-api.R
\name{get_data_count}
\alias{get_data_count}
\title{get_data_count}
\usage{
get_data_count(resource)
}
\arguments{
\item{resource}{}
}
\value{
numeric count of data items.
}
\description{
Get the count of data items available through the API, which is useful for
estimating the maximum number of items available.
}
\examples{
  get_data_count("avoin_yliopisto")

}
