% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipunen-api.R
\name{print.vipunen_api}
\alias{print.vipunen_api}
\title{Print method for vipunen_api class}
\usage{
\method{print}{vipunen_api}(x, ...)
}
\arguments{
\item{x}{vipunen_api object}

\item{...}{ignored}
}
\description{
Print method for vipunen_api class
}
