% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipunen-api.R
\name{vipunen_api}
\alias{vipunen_api}
\title{vipunen_api}
\usage{
vipunen_api(path)
}
\arguments{
\item{path}{character url to be appended to the host.}
}
\value{
vipunen_api (S3) object with the following attributes:
\describe{
\item{content}{a list of parsed JSON content.}
\item{path}{path provided to get the resonse.}
\item{response}{the original response object.}
}
}
\description{
Make a request to one of the Vipunen API's endopoints. The base url is
http://api.vipunen.fi to which a specific url defined by \code{path} is
appended.
}
\details{
This is a low-level function intended to be used by other higher level
functions in the package.
}
